% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{play_button}
\alias{play_button}
\title{Play midi file}
\usage{
play_button(audiofile, fileext = "wav")
}
\arguments{
\item{audiofile}{audiofile}

\item{fileext}{file extension}
}
\value{
html tag of a playbutton
}
\description{
Transform the midi file to fileext format and provide a play button for html
documents.
}
\examples{
\dontrun{
midifile <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
audiofile <- "test.wav"
synthesize_midi(midifile, audiofile)
convert_to_mp3(audiofile)
audiofile \%>\% play_button()
}
}
