% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{synthesize_midi}
\alias{synthesize_midi}
\title{Synthesize midi file}
\usage{
synthesize_midi(midifile, audiofile, soundfont = NULL)
}
\arguments{
\item{midifile}{midi file path}

\item{audiofile}{audio wav file}

\item{soundfont}{path to sf2 sound font}
}
\value{

}
\description{
Needs fluidsynth installed; no idea how this works on windows.
}
\examples{
\dontrun{
midifile <- system.file("extdata", "test_midi_file.mid", package = "raudiomate")
audiofile <- "test.wav"
synthesize_midi(midifile, audiofile)
}
}
