% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{synthesize_midi}
\alias{synthesize_midi}
\title{Synthesize midi file}
\usage{
synthesize_midi(midifile, audiofile, soundfont = NULL, verbose = FALSE)
}
\arguments{
\item{midifile}{path to midi file path (character string)}

\item{audiofile}{path to audio wav file (character string)}

\item{soundfont}{path to sf2 sound font (character string)}

\item{verbose}{logical whether to print fluidsynth command line output; defaults to FALSE}
}
\description{
Generates a wav \code{audiofile} from a \code{midifile} using the \code{soundfont} if specified
(otherwise fluidsynth's default soundfont is used). Needs fluidsynth
to be installed.
}
\examples{
\dontrun{
midifile <- system.file("extdata", "test_midi_file.mid", package = "raudiomate")
audiofile <- "test.wav"
synthesize_midi(midifile, audiofile)
}
}
